+++ example/bin/services/service.dart
@@ -1,0 +1,9 @@
+import 'package:injectable/injectable.dart';
+
+@Injectable(as: Service)
+class DemoService implements Service {
+  @factoryMethod
+  static Future<DemoService> init() async => DemoService();
+}
+
+abstract class Service {}