+++ injectable_generator/lib/dependency_resolver.dart
@@ -10,6 +10,8 @@ import 'injectable_types.dart';
 
 const TypeChecker namedChecker = TypeChecker.fromRuntime(Named);
 const TypeChecker singletonChecker = TypeChecker.fromRuntime(Singleton);
+const TypeChecker injectableChecker = TypeChecker.fromRuntime(Injectable);
+
 const TypeChecker envChecker = TypeChecker.fromRuntime(Environment);
 const TypeChecker registerAsChecker = TypeChecker.fromRuntime(RegisterAs);
 const TypeChecker preResolveChecker = TypeChecker.fromRuntime(PreResolve);
@@ -48,30 +50,37 @@ class DependencyResolver {
       }
     }
 
-    throwBoxedIf(returnType.element is! ClassElement,
-        '${returnType.getDisplayString()} is not a class element');
+    throwIf(
+      returnType.element is! ClassElement,
+      '${returnType.getDisplayString()} is not a class element',
+      element: returnType.element,
+    );
 
     await _checkForParameterizedTypes(returnType);
-    final registerModuleItem = RegisterModuleItem();
-    registerModuleItem.moduleName = moduleClazz.name;
-    registerModuleItem.import = getImport(moduleClazz);
 
+    _dep.moduleName = moduleClazz.name;
+    _dep.initializerName = executableElement.name;
+    _dep.imports.add(getImport(moduleClazz));
+
+    ExecutableElement executableModuleMember;
     if (executableElement is MethodElement) {
-      throwBoxedIf(executableElement.parameters.length > 2,
-          'Error generating [$returnType]! Max number of factory params is 2');
-      registerModuleItem.isMethod = true;
+      _dep.isModuleMethod = true;
 
-      for (var param in executableElement.parameters) {
-        await _resolveAndAddImport(param.type.element);
-        await _checkForParameterizedTypes(param.type);
-        registerModuleItem.params[param.name] = param.type.getDisplayString();
+      if (!executableElement.isAbstract) {
+        executableModuleMember = executableElement;
+      } else {
+        throwIf(
+          executableElement.parameters.isNotEmpty,
+          'Abstract methods can not have injectable or factory paramters',
+          element: executableElement,
+        );
       }
     }
 
     ClassElement clazz;
     if (executableElement.isAbstract) {
       clazz = returnType.element;
-      registerModuleItem.isAbstract = true;
+      _dep.isAbstract = true;
     } else {
       if (returnType.isDartAsyncFuture) {
         final typeArg = returnType as ParameterizedType;
@@ -82,37 +91,68 @@ class DependencyResolver {
         clazz = returnType.element;
       }
     }
-    registerModuleItem.name = executableElement.name;
-    _dep.moduleConfig = registerModuleItem;
     await _resolveAndAddImport(clazz);
-    return _resolveActualType(clazz, typeName);
+    return _resolveActualType(clazz, typeName, executableModuleMember);
   }
 
-  Future<DependencyConfig> _resolveActualType(ClassElement clazz,
-      [String typeName]) async {
+  Future<DependencyConfig> _resolveActualType(
+    ClassElement clazz, [
+    String typeName,
+    ExecutableElement executbaleModuleMemeber,
+  ]) async {
     _dep.type = typeName ?? clazz.name;
-    _dep.bindTo = typeName ?? clazz.name;
+    _dep.typeImpl = typeName ?? clazz.name;
 
+    final injectableAnnotation = injectableChecker
+        .firstAnnotationOf(_annotatedElement, throwOnUnresolved: false);
+
+    var abstractType;
     var inlineEnv;
-    final registerAsAnnotation =
-        registerAsChecker.firstAnnotationOf(_annotatedElement);
-    if (registerAsAnnotation != null) {
-      ConstantReader registerAsReader = ConstantReader(registerAsAnnotation);
-      final abstractType = registerAsReader.peek('abstractType').typeValue;
-      final abstractChecker = TypeChecker.fromStatic(abstractType);
 
+    // set default injectable type to factory
+    _dep.injectableType = InjectableType.factory;
+
+    if (injectableAnnotation != null) {
+      final injectable = ConstantReader(injectableAnnotation);
+      if (injectable.instanceOf(TypeChecker.fromRuntime(LazySingleton))) {
+        _dep.injectableType = InjectableType.lazySingleton;
+      } else if (injectable.instanceOf(TypeChecker.fromRuntime(Singleton))) {
+        _dep.injectableType = InjectableType.singleton;
+        _dep.signalsReady = injectable.peek('signalsReady')?.boolValue;
+        _dep.dependsOn = injectable
+            .peek('dependsOn')
+            ?.listValue
+            ?.map<String>((v) => v.toTypeValue().getDisplayString())
+            ?.toList();
+      }
+      abstractType = injectable.peek('as')?.typeValue;
+      inlineEnv = injectable.peek('env')?.stringValue;
+    }
+
+    if (abstractType == null) {
+      final registerAsAnnotation =
+          registerAsChecker.firstAnnotationOf(_annotatedElement);
+      if (registerAsAnnotation != null) {
+        final registerAsReader = ConstantReader(registerAsAnnotation);
+        abstractType = registerAsReader.peek('abstractType').typeValue;
+        inlineEnv = registerAsReader.peek('env')?.stringValue;
+      }
+    }
+
+    if (abstractType != null) {
+      final abstractChecker = TypeChecker.fromStatic(abstractType);
       final abstractSubtype = clazz.allSupertypes.firstWhere(
           (type) => abstractChecker.isExactly(type.element), orElse: () {
-        throwBoxed(
-            '[${clazz.name}] is not a subtype of [${abstractType.getDisplayString()}]');
+        throwError(
+          '[${clazz.name}] is not a subtype of [${abstractType.getDisplayString()}]',
+          element: clazz,
+        );
         return null;
       });
 
       _dep.type = abstractSubtype.getDisplayString();
 
-      inlineEnv = registerAsReader.peek('env')?.stringValue;
-
-      _dep.bindTo = clazz.name;
+      _dep.typeImpl = clazz.name;
       await _resolveAndAddImport(abstractSubtype.element);
       await _checkForParameterizedTypes(abstractSubtype);
     }
@@ -124,6 +164,7 @@ class DependencyResolver {
             ?.toStringValue();
 
     _dep.preResolve = preResolveChecker.hasAnnotationOfExact(_annotatedElement);
+
     final name = namedChecker
         .firstAnnotationOfExact(_annotatedElement)
         ?.getField('name')
@@ -136,46 +177,32 @@ class DependencyResolver {
       }
     }
 
-    final singletonAnnotation = singletonChecker
-        .firstAnnotationOf(_annotatedElement, throwOnUnresolved: false);
-    if (singletonAnnotation != null) {
-      if (singletonAnnotation.getField('_lazy').toBoolValue()) {
-        _dep.injectableType = InjectableType.lazySingleton;
-      } else {
-        _dep.injectableType = InjectableType.singleton;
-        _dep.dependsOn = singletonAnnotation
-            .getField('dependsOn')
-            ?.toListValue()
-            ?.map<String>((v) => v.toTypeValue().getDisplayString())
-            ?.toList();
-      }
+    ExecutableElement excutableInitilizer;
 
-      _dep.signalsReady =
-          singletonAnnotation.getField('signalsReady')?.toBoolValue();
-    } else {
-      _dep.injectableType = InjectableType.factory;
-    }
-
-    ExecutableElement constructor;
-    if (_dep.moduleConfig == null || _dep.moduleConfig.isAbstract) {
+    if (executbaleModuleMemeber != null) {
+      excutableInitilizer = executbaleModuleMemeber;
+    } else if (!_dep.isFromModule || _dep.isAbstract) {
       final possibleFactories = <ExecutableElement>[
         ...clazz.methods.where((m) => m.isStatic),
         ...clazz.constructors
       ];
 
-      constructor = possibleFactories
-          .firstWhere((m) => constructorChecker.hasAnnotationOf(m), orElse: () {
-        throwBoxedIf(clazz.isAbstract,
-            '''[${clazz.name}] is abstract and can not be registered directly!
-                \n if it has a factory or a create method annotate it with @factoryMethod''');
+      excutableInitilizer = possibleFactories.firstWhere(
+          (m) => constructorChecker.hasAnnotationOfExact(m), orElse: () {
+        throwIf(
+          clazz.isAbstract,
+          '''[${clazz.name}] is abstract and can not be registered directly!
+           if it has a factory or a create method annotate it with @factoryMethod''',
+          element: clazz,
+        );
         return clazz.unnamedConstructor;
       });
+      _dep.isAsync = excutableInitilizer.returnType.isDartAsyncFuture;
     }
-    if (constructor != null) {
-      _dep.isAsync = constructor.returnType.isDartAsyncFuture;
-      _dep.constructorName = constructor.name;
 
-      for (ParameterElement param in constructor.parameters) {
+    if (excutableInitilizer != null) {
+      _dep.constructorName = excutableInitilizer.name;
+      for (ParameterElement param in excutableInitilizer.parameters) {
         final namedAnnotation = namedChecker.firstAnnotationOf(param);
         final instanceName = namedAnnotation
                 ?.getField('type')
@@ -188,35 +215,57 @@ class DependencyResolver {
         var typeName = param.type.getDisplayString();
         if (param.type is TypeParameterType) {
           typeName = _typeArgsMap[param.type.getDisplayString()];
-          throwBoxedIf(typeName == null,
-              'Error generating [${clazz.name}]! Can not resolve dependency of type ${param.type.getDisplayString()}');
+          throwIf(
+            typeName == null,
+            'Can not resolve dependency of type ${param.type.getDisplayString()}',
+            element: param,
+          );
         }
 
         _dep.dependencies.add(InjectedDependency(
           type: typeName,
           name: instanceName,
-          isFactoryParam: _dep.moduleConfig == null &&
-              factoryParamChecker.hasAnnotationOf(param),
+          isFactoryParam: factoryParamChecker.hasAnnotationOfExact(param),
           paramName: param.name,
           isPositional: param.isPositional,
         ));
-
-        throwBoxedIf(
-            _dep.dependencies.where((d) => d.isFactoryParam).length > 2,
-            'Error generating [${clazz.name}]! Max number of factory params is 2');
       }
+      final factoryParamsCount =
+          _dep.dependencies.where((d) => d.isFactoryParam).length;
+
+      throwIf(
+        _dep.preResolve && factoryParamsCount != 0,
+        'Factories with params can not be pre-resolved',
+        element: clazz,
+      );
+
+      throwIf(
+        _dep.isAbstract && factoryParamsCount != 0,
+        'Module dependecies with factory params must have custom initilaizers',
+        element: clazz,
+      );
+
+      throwIf(
+        _dep.injectableType != InjectableType.factory &&
+            factoryParamsCount != 0,
+        'only factories can have parameters',
+        element: clazz,
+      );
+
+      throwIf(
+        factoryParamsCount > 2,
+        'Max number of factory params supported by get_it is 2',
+        element: clazz,
+      );
     }
 
-    throwBoxedIf(
-        _dep.injectableType != InjectableType.factory &&
-            (_dep.dependencies.where((d) => d.isFactoryParam).isNotEmpty ||
-                _dep.moduleConfig?.params?.isNotEmpty == true),
-        'Error generating [${clazz.name}]! only factories can have parameters');
     return _dep;
   }
 
   Future<String> _resolveLibImport(Element element) async {
-    if (element.source == null || isCoreDartType(element.source)) {
+    if (element == null ||
+        element.source == null ||
+        isCoreDartType(element.source)) {
       return null;
     }
     //if element from a system library but not from dart:core
@@ -224,12 +273,10 @@ class DependencyResolver {
       return getImport(element);
     }
     final assetId = await _resolver.assetIdForElement(element);
-    final lib = await _resolver.findLibraryByName(assetId.package);
-    if (lib != null) {
-      return getImport(lib);
-    } else {
-      return getImport(element);
-    }
+
+    final toBeImported =
+        await _resolver.findLibraryByName(assetId.package) ?? element;
+    return getImport(toBeImported);
   }
 
   Future<void> _checkForParameterizedTypes(DartType paramType) async {