+++ injectable_generator/lib/generator/config_code_generator.dart
@@ -2,14 +2,12 @@ import 'dart:async';
 
 import 'package:injectable_generator/dependency_config.dart';
 import 'package:injectable_generator/generator/factory_param_generator.dart';
-import 'package:injectable_generator/generator/lazy_singleton_generator.dart';
 import 'package:injectable_generator/generator/module_factory_generator.dart';
 import 'package:injectable_generator/generator/singleton_generator.dart';
 import 'package:injectable_generator/injectable_types.dart';
 import 'package:injectable_generator/utils.dart';
 
-import 'factory_generator.dart';
-import 'module_factory_param_generator.dart';
+import 'lazy_factory_generator.dart';
 
 // holds all used var names
 // to make sure we don't have duplicate var names
@@ -23,6 +21,7 @@ class ConfigCodeGenerator {
   ConfigCodeGenerator(this.allDeps);
 
   _write(Object o) => _buffer.write(o);
+
   _writeln(Object o) => _buffer.writeln(o);
 
   // generate configuration function from dependency configs
@@ -37,10 +36,8 @@ class ConfigCodeGenerator {
     final Set<DependencyConfig> sorted = {};
     _sortByDependents(allDeps.toSet(), sorted);
 
-    final modules = sorted
-        .where((d) => d.moduleConfig != null)
-        .map((d) => d.moduleConfig.moduleName)
-        .toSet();
+    final modules =
+        sorted.where((d) => d.isFromModule).map((d) => d.moduleName).toSet();
 
     final Set<DependencyConfig> eagerDeps = sorted
         .where((d) => d.injectableType == InjectableType.singleton)
@@ -49,8 +46,8 @@ class ConfigCodeGenerator {
     final lazyDeps = sorted.difference(eagerDeps);
 
     // generate import
-    final imports =
-        sorted.fold<Set<String>>({}, (a, b) => a..addAll(b.allImports));
+    final imports = sorted.fold<Set<String>>(
+        {}, (a, b) => a..addAll(b.imports.where((i) => i != null)));
 
     // add getIt import statement
     imports.add("package:get_it/get_it.dart");
@@ -122,16 +119,13 @@ class ConfigCodeGenerator {
   void _generateDeps(Set<DependencyConfig> deps) {
     deps.forEach((dep) {
       if (dep.injectableType == InjectableType.factory) {
-        if (dep.moduleConfig == null &&
-            dep.dependencies.where((d) => d.isFactoryParam).isNotEmpty) {
+        if (dep.dependencies.any((d) => d.isFactoryParam)) {
           _writeln(FactoryParamGenerator().generate(dep));
-        } else if (dep.moduleConfig?.params?.isNotEmpty == true) {
-          _writeln(ModuleFactoryParamGenerator().generate(dep));
         } else {
-          _writeln(FactoryGenerator().generate(dep));
+          _writeln(LazyFactoryGenerator().generate(dep));
         }
       } else if (dep.injectableType == InjectableType.lazySingleton) {
-        _writeln(LazySingletonGenerator().generate(dep));
+        _writeln(LazyFactoryGenerator(isLazySingleton: true).generate(dep));
       }
     });
   }
@@ -171,10 +165,8 @@ class ConfigCodeGenerator {
 
   Iterable<DependencyConfig> _getAbstractModuleDeps(
       Set<DependencyConfig> deps, String m) {
-    return deps.where((d) =>
-        d.moduleConfig != null &&
-        d.moduleConfig.moduleName == m &&
-        d.moduleConfig.isAbstract);
+    return deps
+        .where((d) => d.isFromModule && d.moduleName == m && d.isAbstract);
   }
 
   void _generateModuleItems(List<DependencyConfig> moduleDeps) {