+++ injectable_generator/lib/generator/factory_param_generator.dart
@@ -4,9 +4,7 @@ import 'package:injectable_generator/generator/register_func_generator.dart';
 class FactoryParamGenerator extends RegisterFuncGenerator {
   @override
   String generate(DependencyConfig dep) {
-    final constructBody = dep.moduleConfig == null
-        ? generateConstructor(dep)
-        : generateConstructorForModule(dep);
+    final initializer = generateInitializer(dep);
 
     var asyncStr = dep.isAsync && !dep.preResolve ? 'Async' : '';
 
@@ -17,11 +15,11 @@ class FactoryParamGenerator extends RegisterFuncGenerator {
       typeArgs['_'] = 'dynamic';
     }
 
-    final argsDeclaration = '<${dep.type},${typeArgs.values.join(',')} >';
-    final methodParams = typeArgs.keys.join(',');
+    final argsDeclaration = '<${dep.type},${typeArgs.values.join(',')}>';
+    final methodParams = typeArgs.keys.join(', ');
 
-    writeln(
-        "g.registerFactoryParam$asyncStr$argsDeclaration(($methodParams)=> $constructBody");
+    write(
+        "g.registerFactoryParam$asyncStr$argsDeclaration(($methodParams)=> $initializer");
 
     closeRegisterFunc(dep);
     return buffer.toString();