+++ injectable_generator/lib/generator/register_func_generator.dart
@@ -1,3 +1,4 @@
+import 'package:injectable/injectable.dart';
 import 'package:injectable_generator/dependency_config.dart';
 import 'package:injectable_generator/generator/config_code_generator.dart';
 import 'package:injectable_generator/utils.dart';
@@ -8,29 +9,55 @@ abstract class RegisterFuncGenerator {
   write(Object o) => buffer.write(o);
   writeln(Object o) => buffer.writeln(o);
   String generate(DependencyConfig dep);
-  String generateConstructor(DependencyConfig dep, {String getIt = 'g'}) {
-    final params = dep.dependencies.map((injectedDep) {
+
+  String generateInitializer(DependencyConfig dep, {String getIt = 'g'}) {
+    final flattenedParams = flattenParams(dep.dependencies, getIt);
+
+    if (dep.isFromModule) {
+      final moduleName = toCamelCase(dep.moduleName);
+      if (!dep.isModuleMethod) {
+        return '$moduleName.${dep.initializerName}';
+      } else {
+        if (dep.isAbstract) {
+          return '$moduleName.${dep.initializerName}()';
+        } else {
+          return '${moduleName}.${dep.initializerName}($flattenedParams)';
+        }
+      }
+    }
+
+    final typeName = stripGenericTypes(dep.typeImpl);
+    final constructorName =
+        dep.constructorName != null && dep.constructorName.isNotEmpty
+            ? '.${dep.constructorName}'
+            : '';
+
+    return '${typeName}$constructorName($flattenedParams)';
+  }
+
+  String flattenParams(List<InjectedDependency> deps, String getIt) {
+    final params = deps.map((injectedDep) {
       var type = injectedDep.type == 'dynamic' ? '' : '<${injectedDep.type}>';
       var instanceName = '';
+
       if (injectedDep.name != null) {
-        instanceName = "instanceName:'${injectedDep.name}'";
+        instanceName = "instanceName: '${injectedDep.name}'";
       }
+
       final paramName =
-          (!injectedDep.isPositional) ? '${injectedDep.paramName}:' : '';
+      (!injectedDep.isPositional) ? '${injectedDep.paramName}: ' : '';
 
       if (injectedDep.isFactoryParam) {
         return '$paramName${injectedDep.paramName}';
       } else {
-        return '${paramName}$getIt$type($instanceName)';
+        return '$paramName$getIt$type($instanceName)';
       }
     }).toList();
 
-    final constructName =
-        dep.constructorName.isEmpty ? "" : ".${dep.constructorName}";
     if (params.length > 2) {
       params.add('');
     }
-    return '${stripGenericTypes(dep.bindTo)}$constructName(${params.join(',')})';
+    return params.join(', ');
   }
 
   String generateAwaitSetup(DependencyConfig dep, String constructBody) {
@@ -45,26 +72,12 @@ abstract class RegisterFuncGenerator {
     return awaitedVar;
   }
 
-  String generateConstructorForModule(DependencyConfig dep) {
-    final mConfig = dep.moduleConfig;
-    final mName = toCamelCase(mConfig.moduleName);
-
-    var initializer = StringBuffer()..write(mConfig.name);
-    if (mConfig.isMethod) {
-      initializer.write('(');
-      initializer.write(mConfig.params.keys.join(','));
-      initializer.write(')');
-    }
-
-    return '$mName.${initializer.toString()}';
-  }
-
   void closeRegisterFunc(DependencyConfig dep) {
     if (dep.signalsReady != null) {
-      write(',signalsReady: ${dep.signalsReady}');
+      write(', signalsReady: ${dep.signalsReady}');
     }
     if (dep.instanceName != null) {
-      write(",instanceName: '${dep.instanceName}'");
+      write(", instanceName: '${dep.instanceName}'");
     }
     write(");");
   }