+++ injectable_generator/lib/generator/singleton_generator.dart
@@ -4,28 +4,26 @@ import 'package:injectable_generator/generator/register_func_generator.dart';
 class SingletonGenerator extends RegisterFuncGenerator {
   @override
   String generate(DependencyConfig dep) {
-    final constructBody = dep.moduleConfig == null
-        ? generateConstructor(dep)
-        : generateConstructorForModule(dep);
+    final initializer = generateInitializer(dep);
 
-    var constructor = constructBody;
+    var constructor = initializer;
     if (dep.registerAsInstance) {
-      constructor = generateAwaitSetup(dep, constructBody);
+      constructor = generateAwaitSetup(dep, initializer);
     }
 
     final typeArg = '<${dep.type}>';
 
     if (dep.isAsync && !dep.preResolve) {
-      writeln('g.registerSingletonAsync$typeArg(()=> $constructor');
+      write('g.registerSingletonAsync$typeArg(()=> $constructor');
       if (dep.dependsOn.isNotEmpty) {
-        write(', dependsOn:${dep.dependsOn}');
+        write(', dependsOn: ${dep.dependsOn}');
       }
     } else {
       if (dep.dependsOn.isEmpty) {
-        writeln("g.registerSingleton$typeArg($constructor");
+        write("g.registerSingleton$typeArg($constructor");
       } else {
-        writeln(
-            'g.registerSingletonWithDependencies$typeArg(()=> $constructor, dependsOn:${dep.dependsOn}');
+        write(
+            'g.registerSingletonWithDependencies$typeArg(()=> $constructor, dependsOn: ${dep.dependsOn}');
       }
     }