+++ injectable_generator/lib/utils.dart
@@ -1,6 +1,7 @@
 // general utils
 import 'package:analyzer/dart/element/element.dart';
 import 'package:analyzer/src/generated/source.dart';
+import 'package:source_gen/source_gen.dart';
 
 // Checks if source is from dart:core library
 bool isCoreDartType(Source source) {
@@ -19,7 +20,7 @@ String getImport(Element element) {
 
   // we don't need to import core dart types
   if (!isCoreDartType(source)) {
-    final path = source.uri.toString();
+    final path = resolveAssetUri(source.uri).toString();
     return path;
   }
   return null;
@@ -44,9 +45,24 @@ void throwBoxed(String message) {
   throw ("\n${pre.padRight(72, '-')}\n$message\n${''.padRight(72, '-')} \n");
 }
 
-void throwBoxedIf(bool condition, String message) {
-  if (condition) {
-    throwBoxed(message);
+void throwSourceError(String message) {
+  final pre = 'Injectable Generator ';
+  throw ("\n${pre.padRight(72, '-')}\n$message\n${''.padRight(72, '-')} \n");
+}
+
+void throwError(String message, {Element element}) {
+  throw InvalidGenerationSourceError(
+    message,
+    element: element,
+  );
+}
+
+void throwIf(bool condition, String message, {Element element}) {
+  if (condition == true) {
+    throw InvalidGenerationSourceError(
+      message,
+      element: element,
+    );
   }
 }
 
@@ -56,3 +72,13 @@ void printBoxed(String message) {
 }
 
 String stripGenericTypes(String type) => RegExp('^([^<]*)').stringMatch(type);
+
+Uri resolveAssetUri(Uri url) => url.scheme == 'asset' &&
+            url.pathSegments.isNotEmpty &&
+            url.pathSegments[1] == 'bin' ||
+        url.pathSegments[1] == 'test'
+    ? url.replace(
+        scheme: '',
+        pathSegments: url.pathSegments.skip(2),
+      )
+    : url;