+++ injectable/CHANGELOG.md
@@ -1,5 +1,15 @@
 # ChangeLog
 
+## [0.4.0] Breaking Changes!
+- Change Methods inside of register modules are treated as factory methods now,
+  so all params are considered injected params unless annotated with @factoryParam.
+- Change registerModule is now replaced with @module
+- Change RegisterAs(Type) Annotation is now replaced with @Injectable(as:Type)
+- Change Singleton.Lazy() is replaced with @LazySingleton()
+- Add generateForDir property to @InjectableInt to specify what directories to generate for.
+- Fix imports issue when working with bin directory
+
+ 
 ## [0.3.0] Breaking Changes!
 - add support for GetIt 4.0.0
 - fix generic types are registered as dynamic